remove nocommit
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1190029 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/src/java/org/apache/lucene/index/codecs/preflex/TermInfosReader.java
@@ -123,8 +123,7 @@ public final class TermInfosReader {
                                                                                    context), fieldInfos, true);
 
         try {
-          // nocommit don't cast to int..
-          index = new TermInfosReaderIndex(indexEnum, indexDivisor, (int) dir.fileLength(indexFileName), totalIndexInterval);
+          index = new TermInfosReaderIndex(indexEnum, indexDivisor, dir.fileLength(indexFileName), totalIndexInterval);
           indexLength = index.length();
         } finally {
           indexEnum.close();
